Implement some menu functions
- Hide (have implemented this manually so we can toggle the menu item)
- Bring to front
- Always On Top
- Quit

I don't know if this belongs here, or if it should be on the Window
or WindowController. However, when it is placed in the WindowController
most of the menu items disable themselves when the window is not in the
foreground which is undesirable when some of the purpose of some of the
menu items is to control whether the window is in the foreground or not.