Convert launcher application to Objective-C to reduce app size
Golly gracious Objective-C is a mess, but it saves about 3MB of
application size due to not requiring multiple copies of the
swift libraries. If there's a way to embed the launcher application
and share the main application's swift libraries we should swap
back over to the swift implementation.

Fix missing attribution for base/source